third_party/blink/web_tests/platform/generic/fast/dom/shadow/parent-tree-scope-in-shadow-expected.txt

Creating shadow dom subtrees from top to bottom.
PASS internals.parentTreeScope(shadow1) is document
PASS internals.parentTreeScope(shadow2) is shadow1
PASS internals.parentTreeScope(shadow3) is shadow2
Creating multiple shadow dom subtrees from bottom to top.
PASS internals.parentTreeScope(shadow1) is document
PASS internals.parentTreeScope(shadow2) is shadow1
PASS internals.parentTreeScope(shadow3) is shadow2
PASS successfullyParsed is true

TEST COMPLETE
